jmetal-problem/src/test/java/org/uma/jmetal/problem/multiobjective/cre/CRE24Test.java
package org.uma.jmetal.problem.multiobjective.cre;

import static org.junit.jupiter.api.Assertions.assertEquals;

import org.junit.jupiter.api.Test;
import org.uma.jmetal.problem.doubleproblem.DoubleProblem;
import org.uma.jmetal.solution.doublesolution.DoubleSolution;

class CRE24Test {
  @Test
  public void shouldConstructorCreateAProblemWithTheRightProperties() {
    DoubleProblem problem = new CRE24();

    assertEquals(7, problem.numberOfVariables());
    assertEquals(2, problem.numberOfObjectives());
    assertEquals(11, problem.numberOfConstraints());
    assertEquals("CRE24", problem.name());
  }

  @Test
  public void shouldEvaluateWorkProperly() {
    DoubleProblem problem = new CRE24();
    DoubleSolution solution = problem.createSolution();
    problem.evaluate(solution);

    assertEquals(7, solution.variables().size());
    assertEquals(2, solution.objectives().length);
    assertEquals(11, solution.constraints().length);
  }
}
